requirements.txt
pafy==0.5.4
psycopg2-binary==2.8.3
PyBluez==0.22
pygame==1.9.6
python-dotenv==0.10.3
python-vlc==3.0.7110
youtube-dl==2019.8.13
